requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile -U
#
absl-py==0.7.1            # via tensorboard, tensorflow-estimator, tensorflow-gpu
astor==0.7.1              # via tensorflow-gpu
certifi==2019.3.9         # via requests
chardet==3.0.4            # via requests
click==7.0
entrypoints==0.3          # via flake8
flake8-docstrings==1.3.0
flake8-polyfill==1.0.2    # via flake8-docstrings
flake8==3.7.7
gast==0.2.2               # via tensorflow-gpu
grpcio==1.19.0            # via tensorboard, tensorflow-gpu
h5py==2.9.0               # via keras-applications
idna==2.8                 # via requests
keras-applications==1.0.7  # via tensorflow-gpu
keras-preprocessing==1.0.9  # via tensorflow-gpu
markdown==3.1             # via tensorboard
mccabe==0.6.1             # via flake8
mock==2.0.0               # via tensorflow-estimator
numpy==1.16.2             # via h5py, keras-applications, keras-preprocessing, pandas, scikit-learn, scipy, tensorboard, tensorflow-estimator, tensorflow-gpu
pandas==0.24.2
pbr==5.1.3                # via mock
protobuf==3.7.1           # via tensorboard, tensorflow-gpu
pycodestyle==2.5.0        # via flake8
pydocstyle==3.0.0         # via flake8-docstrings
pyflakes==2.1.1           # via flake8
python-dateutil==2.8.0    # via pandas
pytz==2019.1              # via pandas
requests==2.21.0
scikit-learn==0.20.3
scipy==1.2.1              # via scikit-learn
six==1.12.0               # via absl-py, grpcio, h5py, keras-preprocessing, mock, protobuf, pydocstyle, python-dateutil, tensorboard, tensorflow-estimator, tensorflow-gpu
snowballstemmer==1.2.1    # via pydocstyle
tensorboard==1.13.1       # via tensorflow-gpu
tensorflow-estimator==1.13.0  # via tensorflow-gpu
tensorflow-gpu==1.13.1
termcolor==1.1.0          # via tensorflow-gpu
urllib3==1.24.1           # via requests
werkzeug==0.15.2          # via tensorboard
wheel==0.33.1             # via tensorboard, tensorflow-gpu
